data_insight/requirements.txt
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
pytest==7.4.0
pytest-cov==4.1.0
pylint==2.17.5
black==23.7.0
langchain==0.0.267
openai==0.27.8
python-dotenv==1.0.0
flask==2.3.3
statsmodels==0.14.0
flask-cors==4.0.0
flask-compress==1.14.0
flask-caching==2.1.0
gunicorn==21.2.0
python-dateutil==2.8.2
pytz==2023.3 